MAGE-1245 Relocate config and base test case classes
