+++ Test/Integration/Indexing/Category/CategoriesIndexingTest.php
@@ -3,7 +3,7 @@
 namespace Algolia\AlgoliaSearch\Test\Integration\Indexing\Category;
 
 use Algolia\AlgoliaSearch\Model\Indexer\Category;
-use Algolia\AlgoliaSearch\Test\Integration\IndexingTestCase;
+use Algolia\AlgoliaSearch\Test\Integration\Indexing\IndexingTestCase;
 
 class CategoriesIndexingTest extends IndexingTestCase
 {
